requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
Pillow==8.0.1
shortuuid==1.0.1
gevent